REST services cleanup, better implementation

+++ src/main/java/com/vasivuk/boardgames/service/impl/ProductServiceImpl.java
@@ -1,5 +1,7 @@
 package com.vasivuk.boardgames.service.impl;
 
+import com.vasivuk.boardgames.exception.EntityAlreadyExistsException;
+import com.vasivuk.boardgames.exception.EntityNotFoundException;
 import com.vasivuk.boardgames.model.Product;
 import com.vasivuk.boardgames.repository.ProductRepository;
 import com.vasivuk.boardgames.service.ProductService;
@@ -8,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -21,23 +24,43 @@ public class ProductServiceImpl implements ProductService {
         return repository.findAll();
     }
 
-    public Product findProductById(Long id) {
-        return null;
-    }
-
-    public Product saveProduct(Product product) {
-        return repository.save(product);
-    }
-
-    public Product updateProduct(Product product) {
-        Product oldProduct = repository.findById(product.getProductId()).get();
-        if(oldProduct == null) {
-            throw new IllegalArgumentException(String.format("Product with id: %s does not exist in system", product.getProductId()));
+    public Product findProductById(Long id) throws EntityNotFoundException {
+        Optional<Product> product = repository.findById(id);
+        if (product.isEmpty()) {
+            throw new EntityNotFoundException("Product doesn't exist.");
         }
-        return repository.save(product);
+        return product.get();
     }
 
-    public void deleteProduct(Long id) {
+    public Product saveProduct(Product newProduct) throws EntityAlreadyExistsException {
+        Optional<Product> product = repository.findByNameIgnoreCase(newProduct.getName());
+        if (product.isPresent()) {
+            throw new EntityAlreadyExistsException("Product with name: " + newProduct.getName() + " already exists");
+        }
+        return repository.save(newProduct);
+    }
+
+    public Product updateProduct(Long id, Product product) throws EntityNotFoundException {
+        Optional<Product> oldProduct = repository.findById(id);
+        if (oldProduct.isEmpty()) {
+            throw new EntityNotFoundException("Product doesn't exist.");
+        }
+        oldProduct.get().setName(product.getName());
+        oldProduct.get().setComplexity(product.getComplexity());
+        oldProduct.get().setDescription(product.getDescription());
+        oldProduct.get().setPrice(product.getPrice());
+        oldProduct.get().setGameTime(product.getGameTime());
+        oldProduct.get().setNumberOfPlayers(product.getNumberOfPlayers());
+        oldProduct.get().setRating(product.getRating());
+
+        return repository.save(oldProduct.get());
+    }
+
+    public void deleteProduct(Long id) throws EntityNotFoundException {
+        Optional<Product> product = repository.findById(id);
+        if (product.isEmpty()) {
+            throw new EntityNotFoundException("Product doesn't exist.");
+        }
         repository.deleteById(id);
     }
 }
